2023 03 13 오류 수정 . 거의 다함

+++ TRPG/MonsterValue.cpp
@@ -1,6 +1,6 @@
 #include "MonsterValue.h"
 //몬스터 초기값
-int monsterHP = 10, monsterAD = 10, monsterDEF = 10;
+double monsterHP = 10, monsterAD = 10, monsterDEF = 10;
